Refactor: Pipeline log vertical slices
