+++ src/Core/Houston.Application/CommandHandlers/PipelineLogCommandHandlers/Get/GetPipelineLogCommand.cs
@@ -1,0 +1,3 @@
+﻿namespace Houston.Application.CommandHandlers.PipelineLogCommandHandlers.Get {
+	public sealed record GetPipelineLogCommand(Guid Id) : IRequest<IResultCommand>;
+}
