+++ src/Core/Houston.Application/CommandHandlers/PipelineLogCommandHandlers/GetAll/GetAllPipelineLogCommand.cs
@@ -1,0 +1,3 @@
+﻿namespace Houston.Application.CommandHandlers.PipelineLogCommandHandlers.GetAll {
+	public sealed record GetAllPipelineLogCommand(Guid PipelineId, int PageSize, int PageIndex) : IRequest<IResultCommand>;
+}
