+++ src/Core/Houston.Application/Usings.cs
@@ -12,6 +12,7 @@ global using Houston.Application.ViewModel;
 global using Houston.Application.ViewModel.ConnectorViewModels;
 global using Houston.Application.ViewModel.ConnectorFunctionViewModels;
 global using Houston.Application.ViewModel.PipelineInstructionViewModels;
+global using Houston.Application.ViewModel.PipelineLogViewModels;
 
 global using Houston.Core.Interfaces.Repository;
 global using Houston.Core.Interfaces.Results;
