+++ src/UserInterface/Houston.API/Controllers/PipelineLogController.cs
@@ -1,22 +1,14 @@
-﻿using AutoMapper;
-using Houston.Application.ViewModel;
-using Houston.Application.ViewModel.ConnectorFunctionViewModels;
-using Houston.Application.ViewModel.PipelineLogViewModels;
-using Houston.Core.Commands.PipelineLogCommands;
-using MediatR;
-using Microsoft.AspNetCore.Mvc;
-using System.Net;
+﻿using Houston.Application.CommandHandlers.PipelineLogCommandHandlers.Get;
+using Houston.Application.CommandHandlers.PipelineLogCommandHandlers.GetAll;
 
 namespace Houston.API.Controllers {
 	[Route("api/[controller]")]
 	[ApiController]
 	public class PipelineLogController : ControllerBase {
 		private readonly IMediator _mediator;
-		private readonly IMapper _mapper;
 
-		public PipelineLogController(IMediator mediator, IMapper mapper) {
+		public PipelineLogController(IMediator mediator) {
 			_mediator = mediator ?? throw new ArgumentNullException(nameof(mediator));
-			_mapper = mapper ?? throw new ArgumentNullException(nameof(mapper));
 		}
 
 		/// <summary>
@@ -26,17 +18,7 @@ namespace Houston.API.Controllers {
 		/// <response code="200">Pipeline log object response</response>
 		/// <response code="404">The requested pipeline log could not be found</response>
 		[HttpGet("item/{id:guid}")]
-		public async Task<IActionResult> Get(Guid id) {
-			var command = new GetPipelineLogCommand(id);
-			var response = await _mediator.Send(command);
-
-			if (response.StatusCode != HttpStatusCode.OK)
-				return StatusCode((int)response.StatusCode, new MessageViewModel(response.ErrorMessage!, response.ErrorCode));
-
-			var view = _mapper.Map<PipelineLogViewModel>(response.Response);
-
-			return Ok(view);
-		}
+		public async Task<IActionResult> Get(Guid id) => await _mediator.Send(new GetPipelineLogCommand(id));
 
 		/// <summary>
 		/// Gets pipeline runs history log
@@ -46,13 +28,6 @@ namespace Houston.API.Controllers {
 		/// <param name="pageIndex"></param>
 		/// <response code="200">Pipeline logs list object response</response>
 		[HttpGet("{pipelineId:guid}")]
-		public async Task<IActionResult> GetAll(Guid pipelineId, [FromQuery] int pageSize = 10, [FromQuery] int pageIndex = 0) {
-			var command = new GetAllPipelineLogCommand(pipelineId, pageSize, pageIndex);
-			var response = await _mediator.Send(command);
-
-			var view = _mapper.Map<List<PipelineLogViewModel>>(response.Response);
-
-			return Ok(new PaginatedItemsViewModel<PipelineLogViewModel>(response.PageIndex, response.PageSize, response.Count, view));
-		}
+		public async Task<IActionResult> GetAll(Guid pipelineId, [FromQuery] int pageSize = 10, [FromQuery] int pageIndex = 0) => await _mediator.Send(new GetAllPipelineLogCommand(pipelineId, pageSize, pageIndex));
 	}
 }
